requirements.txt
flask
gunicorn
cryptography
flask-socketio
eventlet==0.30.2
Flask-Failsafe
mysql-connector-python
python-dotenv
APScheduler
requests